requirements.txt
Flask
Flask-CORS
requests
uWSGI
python-dotenv
